+++ eslint.config.js
@@ -1,0 +1,12 @@
+import js from '@eslint/js';
+import globals from 'globals';
+import { defineConfig } from 'eslint/config';
+
+export default defineConfig([
+    // We assume the linter will only work in the resources/assets directory. Here we will exclude common vender files
+    {
+        ignores: ['**/vendor/*'],
+    },
+    { files: ['**/*.{js,mjs,cjs}'], plugins: { js }, extends: ['js/recommended'] },
+    { files: ['**/*.{js,mjs,cjs}'], languageOptions: { globals: globals.browser } },
+]);