+++ resources/views/components/form/fieldset.blade.php
@@ -15,6 +15,7 @@
                 '[&_[data-slot=label]+[data-slot=description]]:mt-1',
                 '[&_[data-slot=label]+[data-slot=tabs]]:mt-1.5',
                 '[&_[data-slot=hint]+[data-slot=hint]]:mt-1',
+                '[&_[data-slot=error]+[data-slot=error]]:mt-1',
             ])
         }}
     >