+++ src/Fragments/UI/views/livewire/_partials/bookmarks.blade.php
@@ -1,25 +1,26 @@
-@php
-    // $urls = $fragment->urls;
-    // TODO(ben): get fragment urls
-    $urls = [
-        ['url' => '/', 'locale' => 'nl'],
-        ['url' => '/fr', 'locale' => 'fr'],
-        ['url' => '/fr-fr', 'locale' => 'fr-fr'],
-    ];
-@endphp
+<div data-slot="form-group" class="flex flex-wrap items-center gap-2">
+    <x-chief::badge size="sm">#{{ $fragment->bookmark }}</x-chief::badge>
 
-@if (count($urls))
-    <div data-slot="form-group" class="space-y-2">
-        @if (count($urls) == 1)
-            @include('chief-fragments::livewire._partials.bookmark', ['url' => $urls[0]['url'], 'bookmark' => $fragment->bookmark])
-        @elseif (count($urls) > 1)
-            <x-chief::tabs :show-nav="false" :should-listen-for-external-tab="true">
-                @foreach ($urls as $url)
-                    <x-chief::tabs.tab tab-id="{{ $url['locale'] }}">
-                        @include('chief-fragments::livewire._partials.bookmark', ['url' => $url['url'], 'bookmark' => $fragment->bookmark])
-                    </x-chief::tabs.tab>
-                @endforeach
-            </x-chief::tabs>
-        @endif
-    </div>
-@endif
+    <x-chief::button
+        size="xs"
+        x-copy="{
+            content: '#{{ $fragment->bookmark }}',
+            successContent: '#{{ $fragment->bookmark }} gekopieerd!'
+        }"
+    >
+        <x-chief::icon.link />
+        <span>Kopieer</span>
+    </x-chief::button>
+
+    {{--
+        <x-chief::button
+        href="#{{ $fragment->bookmark }}"
+        title="Bekijk dit fragment op de website"
+        target="_blank"
+        rel="noopener"
+        size="xs"
+        >
+        <x-chief::icon.link-square />
+        </x-chief::button>
+    --}}
+</div>