+++ src/Resource/TreeResourceDefault.php
@@ -13,32 +13,35 @@ trait TreeResourceDefault
     public function getTreeModelIds(): array
     {
         $modelClass = static::modelClassName();
+        $modelObject = new $modelClass;
 
-        return DB::table((new $modelClass)->getTable())
+        return DB::table($modelObject->getTable())
             ->orderBy('order')
-            ->select(['id', 'parent_id'])
+            ->select([$modelObject->getKeyName(), 'parent_id'])
             ->get()
             ->all();
     }
 
-    public function getTreeModels(?array $ids = null): Collection
+    public function getTreeModels(?array $ids = null, array $eagerLoading = ['urls', 'tags']): Collection
     {
         $modelClass = static::modelClassName();
         $reflection = (new \ReflectionClass($modelClass));
-        $eagerLoading = [];
+        $modelObject = new $modelClass;
 
-        if ($reflection->implementsInterface(Visitable::class)) {
-            $eagerLoading[] = 'urls';
-        }
+        foreach ($eagerLoading as $relation) {
+            if ($relation == 'urls' && ! $reflection->implementsInterface(Visitable::class)) {
+                unset($eagerLoading[array_search('urls', $eagerLoading)]);
+            }
 
-        if ($reflection->implementsInterface(Taggable::class)) {
-            $eagerLoading[] = 'tags';
+            if ($relation == 'tags' && ! $reflection->implementsInterface(Taggable::class)) {
+                unset($eagerLoading[array_search('tags', $eagerLoading)]);
+            }
         }
 
         $models = $modelClass::withoutGlobalScopes()
             ->with($eagerLoading)
             ->orderBy('order')
-            ->when($ids, fn ($query) => $query->whereIn('id', $ids))
+            ->when($ids, fn ($query) => $query->whereIn($modelObject->getKeyName(), $ids))
             ->get();
 
         // Sort by parent