+++ src/Table/Filters/SelectFilter.php
@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 namespace Thinktomorrow\Chief\Table\Filters;
 
+use Thinktomorrow\Chief\Forms\Fields\Concerns\Select\HasGroupedOptions;
 use Thinktomorrow\Chief\Forms\Fields\Concerns\Select\HasMultiple;
 use Thinktomorrow\Chief\Forms\Fields\Concerns\Select\HasOptions;
 use Thinktomorrow\Chief\Forms\Fields\Concerns\Select\PairOptions;
 
 class SelectFilter extends Filter
 {
+    use HasGroupedOptions;
     use HasMultiple;
     use HasOptions;
 
@@ -18,6 +20,31 @@ class SelectFilter extends Filter
     public function getMultiSelectFieldOptions(?string $locale = null): array
     {
         return PairOptions::convertOptionsToChoices($this->getOptions($locale));
+    }
+
+    public function findLabelByValue(string $value, ?string $locale = null): ?string
+    {
+        $options = $this->getOptions($locale);
+
+        if ($this->hasOptionGroups($locale)) {
+            foreach ($options as $group) {
+                foreach ($group['options'] as $option) {
+                    if ($option['value'] == $value) {
+                        return $option['label'];
+                    }
+                }
+            }
+
+            return null;
+        }
+
+        foreach ($options as $option) {
+            if ($option['value'] == $value) {
+                return $option['label'];
+            }
+        }
+
+        return null;
     }
 
     private function getOptionsCallableParameters(?string $locale = null): array